Refactor CommandLinker exports.
